// src/app/model/Upload.model.ts
export interface UploadLine {
  property1: string;
  property2: string;
  property3: string;
  property4: string;
  property5: string;
  property6: string;
  property7: string;
  property8: string;
  property9: string;
  property10: string;
  property11: string;
  property12: string;
  property13: string;
  property14: string;
  property15: string;
}